Fixes issue 257.
git-svn-id: http://hazelcast.googlecode.com/svn/trunk@1211 3f8e66b6-ca9d-11dd-a2b5-e5f827957e07

+++ hazelcast/src/main/java/com/hazelcast/nio/SerializationHelper.java
@@ -38,7 +38,7 @@ public class SerializationHelper {
             out.writeDouble((Double) obj);
         } else if (obj instanceof Float) {
             out.writeByte(5);
-            out.writeDouble((Float) obj);
+            out.writeFloat((Float) obj);
         } else if (obj instanceof Boolean) {
             out.writeByte(6);
             out.writeBoolean((Boolean) obj);
@@ -66,7 +66,7 @@ public class SerializationHelper {
         if (type == 0) {
             return null;
         } else if (type == 1) {
-            return in.readInt();
+            return in.readLong();
         } else if (type == 2) {
             return in.readInt();
         } else if (type == 3) {
